🔧 test(tools): extractID 挪到跨平台文件,修 Windows 测试构建
command_test.go 是跨平台的(RunCommand 测试运行时按 runtime.GOOS 跳过 Windows),
但它引用的 extractID 定义在 bg_test.go(//go:build !windows)里 —— Windows 下该
符号未定义,导致整个 tools 测试包在 Windows 上编译失败(交叉编译 #195 时暴露)。

extractID 是纯字符串解析、无平台依赖,挪到不带约束的 testhelpers_test.go,
bg_test.go(!windows)与 command_test.go(跨平台)都能用。command_test.go 一字未改,
保留其原有的运行时 GOOS 跳过设计。GOOS=windows go vet ./tools/ 现通过。
